add curvature and remesh functionality
The marker program now has the ability to display more curved hence
"sharper" regions in red. To display these, press "S" on your keyboard.

Further the marker program allows for remeshing the geometry to 2000
polygons. This is the poly count that is used in the main application.

Lastly a status text in the upper left corner is introduces. It is soley
for the purpose of achieving responsiveness.